+++ Sources/Sentry/Public/SentrySDK.h
@@ -333,6 +333,22 @@ SENTRY_NO_INIT
  */
 + (void)close;
 
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+/**
+ * Start a new continuous profiling session if one is not already running.
+ * @warning Continuous profiling mode is experimental and may still contain bugs.
+ * @seealso https://docs.sentry.io/platforms/apple/profiling/
+ */
++ (void)startProfiler;
+
+/**
+ * Stop a continuous profiling session if there is one ongoing.
+ * @warning Continuous profiling mode is experimental and may still contain bugs.
+ * @seealso https://docs.sentry.io/platforms/apple/profiling/
+ */
++ (void)stopProfiler;
+#endif // SENTRY_TARGET_PROFILING_SUPPORTED
+
 @end
 
 NS_ASSUME_NONNULL_END